lab-6.1-mcp-server/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
websockets==12.0
pydantic==2.6.4
redis==5.0.3
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
kubernetes==29.0.0
PyYAML==6.0.1
pytest==8.1.1
